requirements.txt
streamlit
gradio
openai==0.28.1
sentence-transformers
faiss-cpu
pandas
